refactor: split login and authenticated app shell
